rename variables
