+++ src/models/UserProfile.ts
@@ -1,4 +1,4 @@
-export interface UserProfileData {
+export interface UserProfileDataInterface {
   userDescription: string | null
   favoriteActivities: number[] | null
 }
